Refactor the parse logic for Instruction
