+++ .vm/src/program/instruction/mod.rs
@@ -24,6 +24,7 @@ mod operation;
 pub use operation::*;
 
 mod bytes;
+mod parse;
 
 use crate::{Program, Stack};
 use console::{
@@ -55,6 +56,112 @@ use console::{
     },
     program::{Register, RegisterType},
 };
+
+#[derive(Clone, PartialEq, Eq, Hash)]
+pub enum Instruction<N: Network, A: circuit::Aleo<Network = N>> {
+    /// Compute the absolute value of `first`, checking for overflow, and storing the outcome in `destination`.
+    Abs(Abs<N, A>),
+    /// Compute the absolute value of `first`, wrapping around at the boundary of the type, and storing the outcome in `destination`.
+    AbsWrapped(AbsWrapped<N, A>),
+    /// Adds `first` with `second`, storing the outcome in `destination`.
+    Add(Add<N, A>),
+    /// Adds `first` with `second`, wrapping around at the boundary of the type, and storing the outcome in `destination`.
+    AddWrapped(AddWrapped<N, A>),
+    /// Performs a bitwise `and` operation on `first` and `second`, storing the outcome in `destination`.
+    And(And<N, A>),
+    /// Calls a closure on the operands.
+    Call(Call<N, A>),
+    /// Casts the operands into the declared type.
+    Cast(Cast<N, A>),
+    /// Performs a BHP commitment on inputs of 256-bit chunks.
+    CommitBHP256(CommitBHP256<N, A>),
+    /// Performs a BHP commitment on inputs of 512-bit chunks.
+    CommitBHP512(CommitBHP512<N, A>),
+    /// Performs a BHP commitment on inputs of 768-bit chunks.
+    CommitBHP768(CommitBHP768<N, A>),
+    /// Performs a BHP commitment on inputs of 1024-bit chunks.
+    CommitBHP1024(CommitBHP1024<N, A>),
+    // /// Performs a Pedersen commitment on up to a 64-bit input.
+    // CommitPed64(CommitPed64<N, A>),
+    // /// Performs a Pedersen commitment on up to a 128-bit input.
+    // CommitPed128(CommitPed128<N, A>),
+    // /// Divides `first` by `second`, storing the outcome in `destination`.
+    // Div(Div<N, A>),
+    // /// Divides `first` by `second`, wrapping around at the boundary of the type, and storing the outcome in `destination`.
+    // DivWrapped(DivWrapped<N, A>),
+    /// Doubles `first`, storing the outcome in `destination`.
+    Double(Double<N, A>),
+    /// Computes whether `first` is greater than `second` as a boolean, storing the outcome in `destination`.
+    GreaterThan(GreaterThan<N, A>),
+    /// Computes whether `first` is greater than or equal to `second` as a boolean, storing the outcome in `destination`.
+    GreaterThanOrEqual(GreaterThanOrEqual<N, A>),
+    /// Performs a BHP hash on inputs of 256-bit chunks.
+    HashBHP256(HashBHP256<N, A>),
+    /// Performs a BHP hash on inputs of 512-bit chunks.
+    HashBHP512(HashBHP512<N, A>),
+    /// Performs a BHP hash on inputs of 768-bit chunks.
+    HashBHP768(HashBHP768<N, A>),
+    /// Performs a BHP hash on inputs of 1024-bit chunks.
+    HashBHP1024(HashBHP1024<N, A>),
+    /// Performs a Pedersen hash on up to a 64-bit input.
+    HashPED64(HashPED64<N, A>),
+    /// Performs a Pedersen hash on up to a 128-bit input.
+    HashPED128(HashPED128<N, A>),
+    /// Performs a Poseidon hash with an input rate of 2.
+    HashPSD2(HashPSD2<N, A>),
+    /// Performs a Poseidon hash with an input rate of 4.
+    HashPSD4(HashPSD4<N, A>),
+    /// Performs a Poseidon hash with an input rate of 8.
+    HashPSD8(HashPSD8<N, A>),
+    /// Computes the multiplicative inverse of `first`, storing the outcome in `destination`.
+    Inv(Inv<N, A>),
+    /// Computes whether `first` equals `second` as a boolean, storing the outcome in `destination`.
+    IsEqual(IsEqual<N, A>),
+    /// Computes whether `first` does **not** equals `second` as a boolean, storing the outcome in `destination`.
+    IsNotEqual(IsNotEqual<N, A>),
+    /// Computes whether `first` is less than `second` as a boolean, storing the outcome in `destination`.
+    LessThan(LessThan<N, A>),
+    /// Computes whether `first` is less than or equal to `second` as a boolean, storing the outcome in `destination`.
+    LessThanOrEqual(LessThanOrEqual<N, A>),
+    /// Multiplies `first` with `second`, storing the outcome in `destination`.
+    Mul(Mul<N, A>),
+    /// Multiplies `first` with `second`, wrapping around at the boundary of the type, and storing the outcome in `destination`.
+    MulWrapped(MulWrapped<N, A>),
+    /// Returns `false` if `first` and `second` are true, storing the outcome in `destination`.
+    Nand(Nand<N, A>),
+    /// Negates `first`, storing the outcome in `destination`.
+    Neg(Neg<N, A>),
+    /// Returns `true` if neither `first` nor `second` is `true`, storing the outcome in `destination`.
+    Nor(Nor<N, A>),
+    /// Flips each bit in the representation of `first`, storing the outcome in `destination`.
+    Not(Not<N, A>),
+    /// Performs a bitwise `or` on `first` and `second`, storing the outcome in `destination`.
+    Or(Or<N, A>),
+    /// Raises `first` to the power of `second`, storing the outcome in `destination`.
+    Pow(Pow<N, A>),
+    /// Raises `first` to the power of `second`, wrapping around at the boundary of the type, storing the outcome in `destination`.
+    PowWrapped(PowWrapped<N, A>),
+    /// Shifts `first` left by `second` bits, storing the outcome in `destination`.
+    Shl(Shl<N, A>),
+    /// Shifts `first` left by `second` bits, continuing past the boundary of the type, storing the outcome in `destination`.
+    ShlWrapped(ShlWrapped<N, A>),
+    /// Shifts `first` right by `second` bits, storing the outcome in `destination`.
+    Shr(Shr<N, A>),
+    /// Shifts `first` right by `second` bits, continuing past the boundary of the type, storing the outcome in `destination`.
+    ShrWrapped(ShrWrapped<N, A>),
+    /// Squares 'first', storing the outcome in `destination`.
+    Square(Square<N, A>),
+    /// Compute the square root of 'first', storing the outcome in `destination`.
+    SquareRoot(SquareRoot<N, A>),
+    /// Computes `first - second`, storing the outcome in `destination`.
+    Sub(Sub<N, A>),
+    /// Computes `first - second`, wrapping around at the boundary of the type, and storing the outcome in `destination`.
+    SubWrapped(SubWrapped<N, A>),
+    /// Selects `first`, if `condition` is true, otherwise selects `second`, storing the result in `destination`.
+    Ternary(Ternary<N, A>),
+    /// Performs a bitwise `xor` on `first` and `second`, storing the outcome in `destination`.
+    Xor(Xor<N, A>),
+}
 
 /// Creates a match statement that applies the given operation for each instruction.
 ///
@@ -180,112 +287,6 @@ macro_rules! instruction {
     }};
 }
 
-#[derive(Clone, PartialEq, Eq, Hash)]
-pub enum Instruction<N: Network, A: circuit::Aleo<Network = N>> {
-    /// Compute the absolute value of `first`, checking for overflow, and storing the outcome in `destination`.
-    Abs(Abs<N, A>),
-    /// Compute the absolute value of `first`, wrapping around at the boundary of the type, and storing the outcome in `destination`.
-    AbsWrapped(AbsWrapped<N, A>),
-    /// Adds `first` with `second`, storing the outcome in `destination`.
-    Add(Add<N, A>),
-    /// Adds `first` with `second`, wrapping around at the boundary of the type, and storing the outcome in `destination`.
-    AddWrapped(AddWrapped<N, A>),
-    /// Performs a bitwise `and` operation on `first` and `second`, storing the outcome in `destination`.
-    And(And<N, A>),
-    /// Calls a closure on the operands.
-    Call(Call<N, A>),
-    /// Casts the operands into the declared type.
-    Cast(Cast<N, A>),
-    /// Performs a BHP commitment on inputs of 256-bit chunks.
-    CommitBHP256(CommitBHP256<N, A>),
-    /// Performs a BHP commitment on inputs of 512-bit chunks.
-    CommitBHP512(CommitBHP512<N, A>),
-    /// Performs a BHP commitment on inputs of 768-bit chunks.
-    CommitBHP768(CommitBHP768<N, A>),
-    /// Performs a BHP commitment on inputs of 1024-bit chunks.
-    CommitBHP1024(CommitBHP1024<N, A>),
-    // /// Performs a Pedersen commitment on up to a 64-bit input.
-    // CommitPed64(CommitPed64<N, A>),
-    // /// Performs a Pedersen commitment on up to a 128-bit input.
-    // CommitPed128(CommitPed128<N, A>),
-    // /// Divides `first` by `second`, storing the outcome in `destination`.
-    // Div(Div<N, A>),
-    // /// Divides `first` by `second`, wrapping around at the boundary of the type, and storing the outcome in `destination`.
-    // DivWrapped(DivWrapped<N, A>),
-    /// Doubles `first`, storing the outcome in `destination`.
-    Double(Double<N, A>),
-    /// Computes whether `first` is greater than `second` as a boolean, storing the outcome in `destination`.
-    GreaterThan(GreaterThan<N, A>),
-    /// Computes whether `first` is greater than or equal to `second` as a boolean, storing the outcome in `destination`.
-    GreaterThanOrEqual(GreaterThanOrEqual<N, A>),
-    /// Performs a BHP hash on inputs of 256-bit chunks.
-    HashBHP256(HashBHP256<N, A>),
-    /// Performs a BHP hash on inputs of 512-bit chunks.
-    HashBHP512(HashBHP512<N, A>),
-    /// Performs a BHP hash on inputs of 768-bit chunks.
-    HashBHP768(HashBHP768<N, A>),
-    /// Performs a BHP hash on inputs of 1024-bit chunks.
-    HashBHP1024(HashBHP1024<N, A>),
-    /// Performs a Pedersen hash on up to a 64-bit input.
-    HashPED64(HashPED64<N, A>),
-    /// Performs a Pedersen hash on up to a 128-bit input.
-    HashPED128(HashPED128<N, A>),
-    /// Performs a Poseidon hash with an input rate of 2.
-    HashPSD2(HashPSD2<N, A>),
-    /// Performs a Poseidon hash with an input rate of 4.
-    HashPSD4(HashPSD4<N, A>),
-    /// Performs a Poseidon hash with an input rate of 8.
-    HashPSD8(HashPSD8<N, A>),
-    /// Computes the multiplicative inverse of `first`, storing the outcome in `destination`.
-    Inv(Inv<N, A>),
-    /// Computes whether `first` equals `second` as a boolean, storing the outcome in `destination`.
-    IsEqual(IsEqual<N, A>),
-    /// Computes whether `first` does **not** equals `second` as a boolean, storing the outcome in `destination`.
-    IsNotEqual(IsNotEqual<N, A>),
-    /// Computes whether `first` is less than `second` as a boolean, storing the outcome in `destination`.
-    LessThan(LessThan<N, A>),
-    /// Computes whether `first` is less than or equal to `second` as a boolean, storing the outcome in `destination`.
-    LessThanOrEqual(LessThanOrEqual<N, A>),
-    /// Multiplies `first` with `second`, storing the outcome in `destination`.
-    Mul(Mul<N, A>),
-    /// Multiplies `first` with `second`, wrapping around at the boundary of the type, and storing the outcome in `destination`.
-    MulWrapped(MulWrapped<N, A>),
-    /// Returns `false` if `first` and `second` are true, storing the outcome in `destination`.
-    Nand(Nand<N, A>),
-    /// Negates `first`, storing the outcome in `destination`.
-    Neg(Neg<N, A>),
-    /// Returns `true` if neither `first` nor `second` is `true`, storing the outcome in `destination`.
-    Nor(Nor<N, A>),
-    /// Flips each bit in the representation of `first`, storing the outcome in `destination`.
-    Not(Not<N, A>),
-    /// Performs a bitwise `or` on `first` and `second`, storing the outcome in `destination`.
-    Or(Or<N, A>),
-    /// Raises `first` to the power of `second`, storing the outcome in `destination`.
-    Pow(Pow<N, A>),
-    /// Raises `first` to the power of `second`, wrapping around at the boundary of the type, storing the outcome in `destination`.
-    PowWrapped(PowWrapped<N, A>),
-    /// Shifts `first` left by `second` bits, storing the outcome in `destination`.
-    Shl(Shl<N, A>),
-    /// Shifts `first` left by `second` bits, continuing past the boundary of the type, storing the outcome in `destination`.
-    ShlWrapped(ShlWrapped<N, A>),
-    /// Shifts `first` right by `second` bits, storing the outcome in `destination`.
-    Shr(Shr<N, A>),
-    /// Shifts `first` right by `second` bits, continuing past the boundary of the type, storing the outcome in `destination`.
-    ShrWrapped(ShrWrapped<N, A>),
-    /// Squares 'first', storing the outcome in `destination`.
-    Square(Square<N, A>),
-    /// Compute the square root of 'first', storing the outcome in `destination`.
-    SquareRoot(SquareRoot<N, A>),
-    /// Computes `first - second`, storing the outcome in `destination`.
-    Sub(Sub<N, A>),
-    /// Computes `first - second`, wrapping around at the boundary of the type, and storing the outcome in `destination`.
-    SubWrapped(SubWrapped<N, A>),
-    /// Selects `first`, if `condition` is true, otherwise selects `second`, storing the result in `destination`.
-    Ternary(Ternary<N, A>),
-    /// Performs a bitwise `xor` on `first` and `second`, storing the outcome in `destination`.
-    Xor(Xor<N, A>),
-}
-
 /// Derives `From<Operation>` for the instruction.
 ///
 /// ## Example
@@ -317,9 +318,7 @@ macro_rules! opcodes {
 impl<N: Network, A: circuit::Aleo<Network = N>> Instruction<N, A> {
     /// The list of all instruction opcodes.
     pub const OPCODES: &'static [Opcode] = &instruction!(opcodes, Instruction, |None| {});
-}
 
-impl<N: Network, A: circuit::Aleo<Network = N>> Instruction<N, A> {
     /// Returns the opcode of the instruction.
     #[inline]
     pub const fn opcode(&self) -> Opcode {
@@ -358,74 +357,6 @@ impl<N: Network, A: circuit::Aleo<Network = N>> Instruction<N, A> {
         input_types: &[RegisterType<N>],
     ) -> Result<Vec<RegisterType<N>>> {
         instruction!(self, |instruction| instruction.output_types(program, input_types))
-    }
-}
-
-impl<N: Network, A: circuit::Aleo<Network = N>> Parser for Instruction<N, A> {
-    /// Parses a string into an instruction.
-    #[inline]
-    fn parse(string: &str) -> ParserResult<Self> {
-        /// Create an alt parser that matches the instruction.
-        ///
-        /// `nom` documentation notes that alt supports a maximum of 21 parsers.
-        /// The documentation suggests to nest alt to support more parsers, as we do here.
-        /// Note that order of the individual parsers matters.
-        macro_rules! alt_parser {
-            ($v0:expr) => {{ alt(($v0,)) }};
-            ($v0:expr, $v1:expr) => {{ alt(($v0, $v1,)) }};
-            ($v0:expr, $v1:expr, $v2:expr) => {{ alt(($v0, $v1, $v2,)) }};
-            ($v0:expr, $v1:expr, $v2:expr, $v3:expr) => {{ alt(($v0, $v1, $v2, $v3,)) }};
-            ($v0:expr, $v1:expr, $v2:expr, $v3:expr, $v4:expr) => {{ alt(($v0, $v1, $v2, $v3, $v4,)) }};
-            ($v0:expr, $v1:expr, $v2:expr, $v3:expr, $v4:expr, $v5:expr) => {{ alt(($v0, $v1, $v2, $v3, $v4, $v5,)) }};
-            ($v0:expr, $v1:expr, $v2:expr, $v3:expr, $v4:expr, $v5:expr, $v6:expr) => {{ alt(($v0, $v1, $v2, $v3, $v4, $v5, $v6,)) }};
-            ($v0:expr, $v1:expr, $v2:expr, $v3:expr, $v4:expr, $v5:expr, $v6:expr, $v7:expr) => {{ alt(($v0, $v1, $v2, $v3, $v4, $v5, $v6, $v7,)) }};
-            ($v0:expr, $v1:expr, $v2:expr, $v3:expr, $v4:expr, $v5:expr, $v6:expr, $v7:expr, $v8:expr) => {{ alt(($v0, $v1, $v2, $v3, $v4, $v5, $v6, $v7, $v8,)) }};
-            ($v0:expr, $v1:expr, $v2:expr, $v3:expr, $v4:expr, $v5:expr, $v6:expr, $v7:expr, $v8:expr, $v9:expr) => {{ alt(($v0, $v1, $v2, $v3, $v4, $v5, $v6, $v7, $v8, $v9,)) }};
-            ($v0:expr, $v1:expr, $v2:expr, $v3:expr, $v4:expr, $v5:expr, $v6:expr, $v7:expr, $v8:expr, $v9:expr, $( $variants:expr ),*) => {{ alt((
-                alt_parser!($( $variants ),*), $v0, $v1, $v2, $v3, $v4, $v5, $v6, $v7, $v8, $v9,
-            )) }};
-        }
-
-        /// Creates a parser for the given instructions.
-        ///
-        /// ## Example
-        /// ```ignore
-        /// instruction_parsers!(self, |_instruction| {}, { Add, Sub, Mul, Div })
-        /// ```
-        macro_rules! instruction_parsers {
-            ($object:expr, |_instruction| $_operation:block, { $( $variant:ident, )+ }) => {{
-                alt_parser!( $( map($variant::parse, Into::into) ),+ )
-            }};
-        }
-
-        // Parse the whitespace and comments from the string.
-        let (string, _) = Sanitizer::parse(string)?;
-        // Parse the instruction from the string.
-        let (string, instruction) = instruction!(instruction_parsers!(self, _instruction))(string)?;
-        // Parse the whitespace from the string.
-        let (string, _) = Sanitizer::parse_whitespaces(string)?;
-        // Parse the semicolon from the string.
-        let (string, _) = tag(";")(string)?;
-
-        Ok((string, instruction))
-    }
-}
-
-impl<N: Network, A: circuit::Aleo<Network = N>> FromStr for Instruction<N, A> {
-    type Err = Error;
-
-    /// Parses a string into an instruction.
-    #[inline]
-    fn from_str(string: &str) -> Result<Self> {
-        match Self::parse(string) {
-            Ok((remainder, object)) => {
-                // Ensure the remainder is empty.
-                ensure!(remainder.is_empty(), "Failed to parse string. Found invalid character in: \"{remainder}\"");
-                // Return the object.
-                Ok(object)
-            }
-            Err(error) => bail!("Failed to parse string. {error}"),
-        }
     }
 }
