+++ .vm/src/program/instruction/parse.rs
@@ -1,0 +1,104 @@
+// Copyright (C) 2019-2022 Aleo Systems Inc.
+// This file is part of the snarkVM library.
+
+// The snarkVM library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// The snarkVM library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
+
+use super::*;
+
+impl<N: Network, A: circuit::Aleo<Network = N>> Parser for Instruction<N, A> {
+    /// Parses a string into an instruction.
+    #[inline]
+    fn parse(string: &str) -> ParserResult<Self> {
+        /// Create an alt parser that matches the instruction.
+        ///
+        /// `nom` documentation notes that alt supports a maximum of 21 parsers.
+        /// The documentation suggests to nest alt to support more parsers, as we do here.
+        /// Note that order of the individual parsers matters.
+        macro_rules! alt_parser {
+            ($v0:expr) => {{ alt(($v0,)) }};
+            ($v0:expr, $v1:expr) => {{ alt(($v0, $v1,)) }};
+            ($v0:expr, $v1:expr, $v2:expr) => {{ alt(($v0, $v1, $v2,)) }};
+            ($v0:expr, $v1:expr, $v2:expr, $v3:expr) => {{ alt(($v0, $v1, $v2, $v3,)) }};
+            ($v0:expr, $v1:expr, $v2:expr, $v3:expr, $v4:expr) => {{ alt(($v0, $v1, $v2, $v3, $v4,)) }};
+            ($v0:expr, $v1:expr, $v2:expr, $v3:expr, $v4:expr, $v5:expr) => {{ alt(($v0, $v1, $v2, $v3, $v4, $v5,)) }};
+            ($v0:expr, $v1:expr, $v2:expr, $v3:expr, $v4:expr, $v5:expr, $v6:expr) => {{ alt(($v0, $v1, $v2, $v3, $v4, $v5, $v6,)) }};
+            ($v0:expr, $v1:expr, $v2:expr, $v3:expr, $v4:expr, $v5:expr, $v6:expr, $v7:expr) => {{ alt(($v0, $v1, $v2, $v3, $v4, $v5, $v6, $v7,)) }};
+            ($v0:expr, $v1:expr, $v2:expr, $v3:expr, $v4:expr, $v5:expr, $v6:expr, $v7:expr, $v8:expr) => {{ alt(($v0, $v1, $v2, $v3, $v4, $v5, $v6, $v7, $v8,)) }};
+            ($v0:expr, $v1:expr, $v2:expr, $v3:expr, $v4:expr, $v5:expr, $v6:expr, $v7:expr, $v8:expr, $v9:expr) => {{ alt(($v0, $v1, $v2, $v3, $v4, $v5, $v6, $v7, $v8, $v9,)) }};
+            ($v0:expr, $v1:expr, $v2:expr, $v3:expr, $v4:expr, $v5:expr, $v6:expr, $v7:expr, $v8:expr, $v9:expr, $( $variants:expr ),*) => {{ alt((
+                alt_parser!($( $variants ),*), $v0, $v1, $v2, $v3, $v4, $v5, $v6, $v7, $v8, $v9,
+            )) }};
+        }
+
+        /// Creates a parser for the given instructions.
+        ///
+        /// ## Example
+        /// ```ignore
+        /// instruction_parsers!(self, |_instruction| {}, { Add, Sub, Mul, Div })
+        /// ```
+        macro_rules! instruction_parsers {
+            ($object:expr, |_instruction| $_operation:block, { $( $variant:ident, )+ }) => {{
+                alt_parser!( $( map($variant::parse, Into::into) ),+ )
+            }};
+        }
+
+        // Parse the whitespace and comments from the string.
+        let (string, _) = Sanitizer::parse(string)?;
+        // Parse the instruction from the string.
+        let (string, instruction) = crate::instruction!(instruction_parsers!(self, _instruction))(string)?;
+        // Parse the whitespace from the string.
+        let (string, _) = Sanitizer::parse_whitespaces(string)?;
+        // Parse the semicolon from the string.
+        let (string, _) = tag(";")(string)?;
+
+        Ok((string, instruction))
+    }
+}
+
+impl<N: Network, A: circuit::Aleo<Network = N>> FromStr for Instruction<N, A> {
+    type Err = Error;
+
+    /// Parses a string into an instruction.
+    #[inline]
+    fn from_str(string: &str) -> Result<Self> {
+        match Self::parse(string) {
+            Ok((remainder, object)) => {
+                // Ensure the remainder is empty.
+                ensure!(remainder.is_empty(), "Failed to parse string. Found invalid character in: \"{remainder}\"");
+                // Return the object.
+                Ok(object)
+            }
+            Err(error) => bail!("Failed to parse string. {error}"),
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use circuit::network::AleoV0;
+    use console::network::Testnet3;
+
+    type CurrentNetwork = Testnet3;
+    type CurrentAleo = AleoV0;
+
+    #[test]
+    fn test_parse() -> Result<()> {
+        let instruction = "add r0 r1 into r2;";
+        let (remainder, candidate) = Instruction::<CurrentNetwork, CurrentAleo>::parse(instruction)?;
+        assert_eq!("", remainder);
+        assert_eq!(instruction, candidate.to_string());
+        Ok(())
+    }
+}
