+++ circuit/algorithms/src/bhp/commit.rs
@@ -78,7 +78,7 @@ mod tests {
 
     #[test]
     fn test_commit_constant() -> Result<()> {
-        check_commit::<32, 48>(Mode::Constant, 8158, 0, 0, 0)
+        check_commit::<32, 48>(Mode::Constant, 8166, 0, 0, 0)
     }
 
     #[test]
